JS code for selection sort visualization project.
https://www.khanacademy.org/computing/computer-science/algorithms/sorting-algorithms/p/project-selection-sort-visualizer

Description :- 
In this project, you'll visualize the swapping strategy of selection sort, similar to the screenshot on the right.

To start off with, you'll need to bring in the selection sort algorithm code from the last challenge. Then, come up with an array of data and send that through the algorithm. For each iteration of the algorithm, display the current state of the array on the canvas, and then draw a line or arrow that shows the numbers being swapped.

Once you've successfully displayed the array being sorted, create 3 more arrays and display them as well, in a quadrant on the canvas. Create one array with reverse sorted values, one array where a few values are the same, and one array where only one value is out of order.